Hack described here: https://hydrodesktop.codeplex.com/workitem/8676
Added a method in Program.cs that knows where to look for System.SQLite.dll and
System.SQLite.Linq.dll based on the operating system. This method is subcribed
to the AssemblyResolve event of the AppDomain.

Removed sqlite3.dll which would be necessary for Windows only if we use the same
Sytem.Data.SQLite.dlls that Mono uses. Added back in SQLite.Interop.dll for use
by Windows' System.Data.SQLite.dlls.